Incorporate input variables that update with text input and setup URL Parameters
This commit adds in the input variables and features that allow them to be updated on page loading in addition to click() and change() events. The variables can now update based on URL parameters.

Incorporate required inputs and enable update on page load

This commit adds in the input variables and features that allow them to be updated on page loading in addition to click() and change() events. The variables can now update based on URL parameters